+++ public/atompagebuilder/app/_app/immutable/chunks/2-d730c1f3.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/(app)/_layout.svelte-eee1a421.js";export{t as component};