+++ public/atompagebuilder/app/_app/immutable/chunks/4-28e482b3.js
@@ -1,1 +1,0 @@
-import{default as t}from"../components/pages/(app)/_page.svelte-c9c9bfdb.js";export{t as component};