+++ public/atompagebuilder/app/_app/immutable/chunks/4-aec43c65.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/(app)/_page.svelte-f53bd645.js";export{t as component};