+++ public/atompagebuilder/app/_app/immutable/chunks/5-308fd34d.js
@@ -1,1 +1,0 @@
-import{default as t}from"../components/pages/(editor)/editor/_page.svelte-9e4f41a6.js";export{t as component};